+++ internal/pkg/ocisif/util.go
@@ -6,18 +6,38 @@
 package ocisif
 
 import (
+	"strings"
+
 	v1 "github.com/google/go-containerregistry/pkg/v1"
+	imagespec "github.com/opencontainers/image-spec/specs-go/v1"
 	ocitsif "github.com/sylabs/oci-tools/pkg/sif"
+	"github.com/sylabs/oci-tools/pkg/sourcesink"
 	"github.com/sylabs/sif/v2/pkg/sif"
 )
 
 // GetSingleImage returns a v1.Image from an OCI-SIF, that must contain a single
-// image.
+// image. It ignores any cosign images that may be present in the OCI-SIF.
 func GetSingleImage(fi *sif.FileImage) (v1.Image, error) {
 	ofi, err := ocitsif.FromFileImage(fi)
 	if err != nil {
 		return nil, err
 	}
 
-	return ofi.Image(nil)
+	return ofi.Image(SkipCosignMatcher)
+}
+
+// SkipCosignMatcher matches all images / indices, except those that are related
+// to cosign images, as annotated using the sylabs/oci-tools ref.name convention.
+func SkipCosignMatcher(d v1.Descriptor) bool {
+	if d.Annotations == nil {
+		return true
+	}
+	refName, ok := d.Annotations[imagespec.AnnotationRefName]
+	if !ok {
+		return true
+	}
+	if strings.HasPrefix(refName, sourcesink.CosignPlaceholderRepo) {
+		return false
+	}
+	return true
 }